+++ client/Packages/com.beamable.server/Editor/BeamServicesCodeWatcher.cs
@@ -40,6 +40,13 @@ namespace Beamable.Server.Editor
 		public override int GetHashCode() => (ServiceName != null ? ServiceName.GetHashCode() : 0);
 	}
 
+	[Serializable]
+	public struct ServiceDependencyChecksum
+	{
+		public string ServiceName;
+		public string Checksum;
+	}
+
 	public enum BeamCodeClass
 	{
 		Invalid,
@@ -56,6 +63,9 @@ namespace Beamable.Server.Editor
 
 		private List<BeamServiceCodeHandle> LatestCodeHandles;
 		private Task CheckSumCalculation;
+
+		private Dictionary<MicroserviceDescriptor, ServiceDependencyChecksum> ServiceToChecksum =
+			new Dictionary<MicroserviceDescriptor, ServiceDependencyChecksum>();
 
 		public void OnInitialized()
 		{
@@ -75,6 +85,13 @@ namespace Beamable.Server.Editor
 				AsmDefInfo = desc.ConvertToInfo(),
 				CodeDirectory = Path.GetDirectoryName(desc.ConvertToInfo().Location),
 			}).ToList();
+
+			ServiceToChecksum =
+				registry.Descriptors.ToDictionary(desc => desc, desc => new ServiceDependencyChecksum
+				{
+					ServiceName = desc.Name,
+					Checksum = DependencyResolver.GetDependencies(desc).GetDependencyChecksum()
+				});
 
 			var storageCodeHandles = registry.StorageDescriptors.Select(desc => new BeamServiceCodeHandle()
 			{
@@ -109,8 +126,12 @@ namespace Beamable.Server.Editor
 			tasks.AddRange(LatestCodeHandles.Select((beamServiceCodeHandle, index) => Task.Factory.StartNew(() =>
 			{
 				var path = beamServiceCodeHandle.CodeDirectory;
-				var files = Directory.GetFiles(path);
-
+				var files = Directory.GetFiles(path)
+				                     .Where(file => !file.EndsWith(".meta"));
+				if (MicroserviceConfiguration.Instance.EnableHotModuleReload)
+				{
+					files = files.Where(file => !file.EndsWith(".cs")).ToArray();
+				}
 				var filesBytes = files.SelectMany(File.ReadAllBytes).ToArray();
 				var md5 = MD5.Create();
 				var checksum = md5.ComputeHash(filesBytes);
@@ -158,7 +179,7 @@ namespace Beamable.Server.Editor
 
 				// If it changes, we need to inform that all services that depend on it must be rebuilt.
 				if (latestCommonCodeHandle.CodeClass != BeamCodeClass.Invalid && serializedCommonCodeHandle.CodeClass != BeamCodeClass.Invalid &&
-					!latestCommonCodeHandle.Equals(serializedCommonCodeHandle.Checksum))
+					!latestCommonCodeHandle.Checksum.Equals(serializedCommonCodeHandle.Checksum))
 				{
 					servicesInNeedOfImageRebuild
 						.AddRange(
@@ -193,11 +214,35 @@ namespace Beamable.Server.Editor
 			}
 
 			// Handle notification of services in need of rebuilds
+			servicesInNeedOfImageRebuild = servicesInNeedOfImageRebuild.Distinct().ToList();
+			if (microserviceConfiguration.EnableHotModuleReload)
 			{
-				servicesInNeedOfImageRebuild = servicesInNeedOfImageRebuild.Distinct().ToList();
+				var registry = BeamEditor.GetReflectionSystem<MicroserviceReflectionCache.Registry>();
+				var dirtyServices = registry.Descriptors.Where(desc =>
+				{
+					if (!ServiceToChecksum.TryGetValue(desc, out var currentDependency))
+					{
+						return true;
+					}
+
+					var existingChecksum = microserviceConfiguration.ServiceDependencyChecksums.FirstOrDefault(
+						service => service.ServiceName == desc.Name);
+
+					return !string.Equals(existingChecksum.Checksum, currentDependency.Checksum);
+				}).ToList();
+
+				microserviceConfiguration.ServiceDependencyChecksums =
+					ServiceToChecksum.Select(kvp => kvp.Value).ToList();
+
+				foreach (var service in dirtyServices)
+				{
+					var _ = RebootContainer(service);
+				}
+			} else
+			{
 				if (servicesInNeedOfImageRebuild.Count > 0)
 				{
-					GlobalStorage.AddOrReplaceHint(BeamHintType.Validation,
+					GlobalStorage.AddOrReplaceHint(BeamHintType.Hint,
 												   BeamHintDomains.BEAM_CSHARP_MICROSERVICES_DOCKER,
 												   BeamHintIds.ID_CHANGES_NOT_DEPLOYED_TO_LOCAL_DOCKER,
 												   servicesInNeedOfImageRebuild);
@@ -209,11 +254,17 @@ namespace Beamable.Server.Editor
 			}
 		}
 
+		private async Promise RebootContainer(MicroserviceDescriptor descriptor)
+		{
+			var model = MicroservicesDataModel.Instance.GetMicroserviceModel(descriptor);
+			await model.Builder.CheckIfIsRunning();
+			if (!model.IsRunning) return;
+			await model.BuildAndRestart();
+		}
+
 		[DidReloadScripts]
 		private static void WatchMicroserviceFiles()
 		{
-			Debug.Log("----------------- DID RELOAD SCRIPTS ---------");
-
 			// If we are not initialized, delay the call until we are.
 			if (!BeamEditor.IsInitialized || !MicroserviceEditor.IsInitialized)
 			{
@@ -231,17 +282,7 @@ namespace Beamable.Server.Editor
 				return;
 			}
 
-			var derp = BeamEditor.GetReflectionSystem<MicroserviceReflectionCache.Registry>();
-			foreach (var service in derp.Descriptors)
-			{
-				BuildUtils.UpdateBuildContextWithSource(service);
-				var client = new MicroserviceAdminClient(service, new PlatformRequester("", null, null));
-				// client.ReloadCode().Then(_ =>
-				// {
-				// 	Debug.Log("CODE HAS RELOADED FOR " + service.Name);
-				// });
-
-			}
+			EditorApplication.quitting += CleanupRunningContainers;
 
 			// Check for the hint regarding local changes that are not deployed to your local docker environment
 			codeWatcher.CheckForLocalChangesNotYetDeployed();
@@ -271,12 +312,9 @@ namespace Beamable.Server.Editor
 					MicroserviceConfiguration.Instance.Microservices.RemoveAll(s => s.ServiceName == serviceName);
 				}
 
-				// For each C#MS that DOES exist, we see if they are new or if there are were changes made to files in their Assembly.
-				var changedOrNewMSCode = latestMSHandles.Where(h => DetectChangesInCodeHandle(h, msHandlesOnPreviousDomainReload)).ToList();
-
-				// For each changed or new handle, gets the descriptor for the C#MS and [re]-generates the client source code.
-				var changedOrNewMsDescriptors = changedOrNewMSCode.Select(h => registry.Descriptors.FirstOrDefault(d => d.Type.Name == h.ServiceName));
-				foreach (var descriptor in changedOrNewMsDescriptors)
+				// For every handle, simply by existing, gets the descriptor for the C#MS and [re]-generates the client source code.
+				var latestDescriptors = latestMSHandles.Select(h => registry.Descriptors.FirstOrDefault(d => d.Type.Name == h.ServiceName));
+				foreach (var descriptor in latestDescriptors)
 				{
 					Assert.IsTrue(descriptor != null, $"You should never see this! The final substring of the Assembly Name must be the same as the ServiceName.");
 					GenerateClientSourceCode(descriptor);
@@ -305,7 +343,42 @@ namespace Beamable.Server.Editor
 			return (!alreadyExisted || checksumDiffers);
 		}
 
-		private static void GenerateClientSourceCode(MicroserviceDescriptor service)
+		public static void CleanupRunningContainers()
+		{
+			try
+			{
+				// this method should exist in some sort of Docker system. But until we have that...
+				var registry = BeamEditor.GetReflectionSystem<MicroserviceReflectionCache.Registry>();
+				foreach (var service in registry.Descriptors)
+				{
+					var generatorDesc = new MicroserviceDescriptor
+					{
+						Name = service.Name + "_generator",
+						AttributePath = service.AttributePath,
+						Type = service.Type
+					};
+
+					var kill = new StopImageCommand(service);
+					var killGenerator = new StopImageCommand(generatorDesc);
+					kill.Start(null);
+					killGenerator.Start(null);
+				}
+
+				foreach (var storage in registry.StorageDescriptors)
+				{
+					var kill = new StopImageCommand(storage);
+					var killTool = new StopImageCommand(storage.LocalToolContainerName);
+					kill.Start(null);
+					killTool.Start(null);
+				}
+			}
+			catch
+			{
+				Debug.LogError("Failed to clean up running docker containers");
+			}
+		}
+
+		public static void GenerateClientSourceCode(MicroserviceDescriptor service, bool force=false)
 		{
 			// create silly descriptor
 			var generatorDesc = new MicroserviceDescriptor
@@ -318,7 +391,7 @@ namespace Beamable.Server.Editor
 			var check = new CheckImageReturnableCommand(generatorDesc);
 			check.Start(null).Then(isRunning =>
 			{
-				if (isRunning) return;
+				if (isRunning && !force) return;
 				// definately stop the image, even if there was doubt it was running. Because if we do a "build", any existing image will ABSOLUTELY be ruined by the overcopy.
 				new StopImageReturnableCommand(generatorDesc).Start(null).Then(__ =>
 				{
@@ -327,19 +400,11 @@ namespace Beamable.Server.Editor
 					{
 						var clientCommand = new RunClientGenerationCommand(generatorDesc);
 						clientCommand.Start();
+						// TODO: add some sort of "cleanup" operation
+						// TODO: consider add info hint when the generator image is running
 					});
 				});
 			});
-
-			// var serviceName = service.Name;
-			// Directory.CreateDirectory("Assets/Beamable/AutoGenerated/Microservices");
-			// var targetFile = $"Assets/Beamable/Autogenerated/Microservices/{serviceName}Client.cs";
-			// var tempFile = Path.Combine("Temp", $"{serviceName}Client.cs");
-			//
-			// var generator = new ClientCodeGenerator(service);
-			// generator.GenerateCSharpCode(tempFile);
-			//
-			// File.Copy(tempFile, targetFile, true);
 		}
 	}
 }