+++ client/Packages/com.beamable.server/Editor/MicroserviceAdminClient.cs
@@ -24,17 +24,6 @@ namespace Beamable.Server.Editor
 			return res.ContainerExists ? $"http://{res.LocalAddress}":null;
 		}
 
-		public async Promise<string> GetCurrentToken()
-		{
-			var addr = await GetAddress();
-			if (string.IsNullOrEmpty(addr))
-			{
-				return "";
-			}
-			var token = await _httpRequester.ManualRequest<string>(Method.GET, addr + "/token", parser: x=> x);
-			return token;
-		}
-
 		public async Promise<bool> RebuildRouteTable()
 		{
 			var addr = await GetAddress();