+++ microservice/microservice/dbmicroservice/LocalDebugService.cs
@@ -21,40 +21,8 @@ namespace Beamable.Server {
             var server = new WebServer(HEALTH_PORT)
                 .WithModule(new ActionModule("/health", HttpVerbs.Any, HealthCheck))
                 .WithAction("/routes/scan", HttpVerbs.Post, RebuildRoutes)
-                .WithAction("/token", HttpVerbs.Get, GetChangeToken)
                 ;
             server.RunAsync();
-        }
-
-        private async Task GetChangeToken(IHttpContext context)
-        {
-            /*
-             * Load up a specific class that
-             */
-            try
-            {
-                var typeName =
-                    _beamableService.MicroserviceType.AssemblyQualifiedName.Replace(
-                        _beamableService.MicroserviceType.Name, "Beamable__Change_Token_Class");
-                var type = Type.GetType(typeName, true);
-
-                var method = type.GetMethod("GetToken", BindingFlags.Static | BindingFlags.Public);
-                var value = method.Invoke(null, new object[] { });
-                var token = value?.ToString() ?? throw new Exception("Invalid token" + value);
-
-                if (!string.IsNullOrEmpty(_lastToken) && !string.Equals(_lastToken, token))
-                {
-                    _beamableService.RebuildRouteTable();
-                }
-
-                _lastToken = token;
-                await context.SendStringAsync(token, "text", Encoding.Default);
-            }
-            catch (Exception ex)
-            {
-                context.Response.StatusCode = 500;
-                await context.SendStringAsync(ex.Message + "\n\n" + ex.StackTrace, "text", Encoding.Default);
-            }
         }
 
         private async Task RebuildRoutes(IHttpContext context)